pyproject.toml
[project]
name = "open-router-key-viewer"
version = "0.3.0"
description = "Desktop viewer for OpenRouter key quota and account credits."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "SunAnICB" },
]
dependencies = [
    "pyside6-fluent-widgets>=1.11.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
    "pytest>=9.0.3",
]
